cap nhat bai tap lam them 1 version 6.0

+++ backend/module_2/src/bai_tap_lam_them_1/service/VehicleService.java
@@ -1,17 +1,18 @@
 package module_2.src.bai_tap_lam_them_1.service;
 
-import module_2.src.bai_tap_lam_them_1.entity.Vehicle;
-import module_2.src.bai_tap_lam_them_1.entity.VehicleType;
+import module_2.src.bai_tap_lam_them_1.dto.VehicleRequest;
+import module_2.src.bai_tap_lam_them_1.entity.*;
 import module_2.src.bai_tap_lam_them_1.repository.IVehicleRepository;
-import module_2.src.bai_tap_lam_them_1.repository.VehicleRepository;
 
 import java.util.List;
 
 public class VehicleService implements IVehicleService {
     private final IVehicleRepository vehicleRepository;
+    private final VehicleValidator vehicleValidator;
 
-    public VehicleService(IVehicleRepository vehicleRepository) {
+    public VehicleService(IVehicleRepository vehicleRepository, VehicleValidator vehicleValidator) {
         this.vehicleRepository = vehicleRepository;
+        this.vehicleValidator = vehicleValidator;
     }
 
     public boolean add(Vehicle vehicle) {
@@ -19,7 +20,6 @@ public class VehicleService implements IVehicleService {
             throw new IllegalArgumentException("Vehicle cannot be null");
         }
         return this.vehicleRepository.add(vehicle);
-
     }
 
     public List<Vehicle> findByLicensePlate(String plate) {
@@ -27,12 +27,46 @@ public class VehicleService implements IVehicleService {
     }
 
     public boolean deleteByLicensePlate(String plate) {
+        if (!this.vehicleValidator.isValidLicensePlate(plate)) {
+            throw new IllegalArgumentException("the license plate is invalid");
+        }
         return this.vehicleRepository.deleteByLicensePlate(plate);
     }
 
-
     public List<Vehicle> getAll() {
         return this.vehicleRepository.getAll();
+    }
+
+    @Override
+    public boolean add(VehicleRequest request) {
+        if (request == null) {
+            throw new IllegalArgumentException("Vehicle cannot be null");
+        }
+
+        if (!this.vehicleValidator.isValidLicensePlate(request.licensePlate)) {
+            throw new IllegalArgumentException("the license plate is invalid");
+        }
+        
+        Vehicle vehicle = VehicleFactory.create(request.vehicleType);
+
+        vehicle.setLicensePlate(request.licensePlate);
+        vehicle.setManufacture(request.manufacture);
+        vehicle.setManufactureYears(request.manufactureYears);
+        vehicle.setOwnerName(request.ownerName);
+
+        switch (request.vehicleType) {
+            case TRUCK:
+                ((Truck) vehicle).setPayloadCapacity(request.payloadCapacity);
+                break;
+            case CAR:
+                ((Car) vehicle).setSeats(request.seats);
+                ((Car) vehicle).setType(request.type);
+                break;
+            case BIKE:
+                ((Bike) vehicle).setEnginePower(request.enginePower);
+                break;
+        }
+        return vehicleRepository.add(vehicle);
     }
 
     public List<Vehicle> getType(VehicleType vehicleType) {
